RichCommand: PermanentlyDelete
