+++ src/Files.App/Strings/en-US/Resources.resw
@@ -3184,4 +3184,10 @@
   <data name="ShowCheckboxesWhenSelectingItems" xml:space="preserve">
     <value>Show checkboxes when selecting items</value>
   </data>
+  <data name="PermanentlyDeleteDescription" xml:space="preserve">
+    <value>Permanently delte item(s)</value>
+  </data>
+  <data name="PermanentlyDeleteDescription" xml:space="preserve">
+    <value>Permanently delte item(s)</value>
+  </data>
 </root>
